+++ examples/resources/morpheus_task_email/resource.tf
@@ -1,7 +1,7 @@
 resource "hpe_morpheus_task_email" "tfexample_email" {
   name                        = "tfexample_email"
   code                        = "tfexample_email"
-  labels                      = ["demo", "terraform"]
+  labels                      = ["demo","terraform"]
   email_address               = "<%=instance.createdByEmail%>"
   subject                     = "<%=instance.hostname%> provisioning complete"
   source                      = "local"
