+++ examples/resources/morpheus_task_email/resource_git.tf
@@ -1,7 +1,7 @@
 resource "hpe_morpheus_task_email" "tfexample_email_git" {
   name                        = "tfexample_email_git"
   code                        = "tfexample_email_git"
-  labels                      = ["demo", "terraform"]
+  labels                      = ["demo","terraform"]
   email_address               = "<%=instance.createdByEmail%>"
   subject                     = "<%=instance.hostname%> provisioning complete"
   source                      = "repository"
