+++ internal/sdkv2/morpheus/resources/task/morpheus_task_email_resource.tf.tmpl
@@ -1,4 +1,4 @@
-resource "hpe_morpheus_task_email" "{{.Name}}" {
+resource "hpe_morpheus_task_email" "tfexample_email" {
   name                        = "{{.Name}}"
   code                        = "{{.Code}}"
   labels                      = {{.Labels}}
