+++ internal/sdkv2/morpheus/resources/task/morpheus_task_email_resource_example.go
@@ -2,6 +2,6 @@
 
 package task
 
-//go:generate go run ../../../../../cmd/render -out examples/resources/morpheus_task_email/resource.tf task_email_resource.tf.tmpl Name tfexample_email Code tfexample_email Labels ["demo","terraform"] EmailAddress "<%=instance.createdByEmail%>" Subject "<%=instance.hostname%> provisioning complete" Source local Content "Your instance <%=instance.hostname%> was provisioned." SkipWrappedEmailTemplate false Retryable true RetryCount 1 RetryDelaySeconds 10 AllowCustomConfig true
-//go:generate go run ../../../../../cmd/render -out examples/resources/morpheus_task_email/resource_url.tf task_email_resource_url.tf.tmpl Name tfexample_email_url Code tfexample_email_url Labels ["demo","terraform"] EmailAddress "<%=instance.createdByEmail%>" Subject "<%=instance.hostname%> provisioning complete" Source url ContentUrl "https://example.com/example.txt" SkipWrappedEmailTemplate false Retryable true RetryCount 1 RetryDelaySeconds 10 AllowCustomConfig true
-//go:generate go run ../../../../../cmd/render -out examples/resources/morpheus_task_email/resource_git.tf task_email_resource_git.tf.tmpl Name tfexample_email_git Code tfexample_email_git Labels ["demo","terraform"] EmailAddress "<%=instance.createdByEmail%>" Subject "<%=instance.hostname%> provisioning complete" Source repository ContentPath "example.txt" RepositoryId 1 VersionRef main SkipWrappedEmailTemplate false Retryable true RetryCount 1 RetryDelaySeconds 10 AllowCustomConfig true
+//go:generate /bin/sh -c "go run ../../../../../cmd/render -out examples/resources/morpheus_morpheus_task_email/resource.tf morpheus_task_email_resource.tf.tmpl Name tfexample_email Code tfexample_email Labels '[\"demo\",\"terraform\"]' EmailAddress '<%=instance.createdByEmail%>' Subject '<%=instance.hostname%> provisioning complete' Source local Content 'Your instance <%=instance.hostname%> was provisioned.' SkipWrappedEmailTemplate false Retryable true RetryCount 1 RetryDelaySeconds 10 AllowCustomConfig true"
+//go:generate /bin/sh -c "go run ../../../../../cmd/render -out examples/resources/morpheus_morpheus_task_email/resource_git.tf morpheus_task_email_resource_git.tf.tmpl Name tfexample_email_git Code tfexample_email_git Labels '[\"demo\",\"terraform\"]' EmailAddress '<%=instance.createdByEmail%>' Subject '<%=instance.hostname%> provisioning complete' Source repository ContentPath example.txt RepositoryId 1 VersionRef main SkipWrappedEmailTemplate false Retryable true RetryCount 1 RetryDelaySeconds 10 AllowCustomConfig true"
+//go:generate /bin/sh -c "go run ../../../../../cmd/render -out examples/resources/morpheus_morpheus_task_email/resource_url.tf morpheus_task_email_resource_url.tf.tmpl Name tfexample_email_url Code tfexample_email_url Labels '[\"demo\",\"terraform\"]' EmailAddress '<%=instance.createdByEmail%>' Subject '<%=instance.hostname%> provisioning complete' Source url ContentUrl https://example.com/example.txt SkipWrappedEmailTemplate false Retryable true RetryCount 1 RetryDelaySeconds 10 AllowCustomConfig true"
