+++ internal/sdkv2/morpheus/resources/task/morpheus_task_email_resource_git.tf.tmpl
@@ -1,4 +1,4 @@
-resource "hpe_morpheus_task_email" "{{.Name}}" {
+resource "hpe_morpheus_task_email" "tfexample_email_git" {
   name                        = "{{.Name}}"
   code                        = "{{.Code}}"
   labels                      = {{.Labels}}
