+++ internal/sdkv2/morpheus/resources/task/morpheus_task_email_resource_test.go
@@ -3,42 +3,23 @@
 package task_test
 
 import (
-	"context"
-	"os"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
-	"github.com/hashicorp/terraform-plugin-mux/tf5to6server"
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 
 	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/testhelpers"
-	sdkv2morpheus "github.com/HPE/terraform-provider-hpe/internal/sdkv2/morpheus"
 )
 
-func TestMain(m *testing.M) {
-	code := m.Run()
-
-	testhelpers.WriteMergedResults()
-
-	os.Exit(code)
-}
-
-func newProviderWithError() (tfprotov6.ProviderServer, error) {
-	return tf5to6server.UpgradeServer(context.Background(), sdkv2morpheus.Provider().GRPCProvider)
-}
-
-var testAccProtoV6ProviderFactories = map[string]func() (
-	tfprotov6.ProviderServer, error,
-){
-	"hpe": newProviderWithError,
-}
-
-func RenderTaskEmailConfig(t *testing.T, overrides map[string]string) (string, error) {
+func RenderMorpheusTaskEmailConfig(
+	t *testing.T,
+	name string,
+	overrides map[string]string,
+) (string, error) {
 	t.Helper()
 
 	defaults := map[string]string{
-		"Name":                     acctest.RandomWithPrefix(t.Name()),
+		"Name":                     name,
 		"Code":                     "tfexample_email",
 		"Labels":                   `["demo","terraform"]`,
 		"EmailAddress":             "<%=instance.createdByEmail%>",
@@ -61,7 +42,8 @@ func RenderTaskEmailConfig(t *testing.T, overrides map[string]string) (string, e
 		args = append(args, key, value)
 	}
 
-	return testhelpers.RenderExample(t, "task_email_resource.tf.tmpl", args...)
+	return testhelpers.RenderExample(t,
+		"morpheus_task_email_resource.tf.tmpl", args...)
 }
 
 func TestAccMorpheusTaskEmailExampleOk(t *testing.T) {
@@ -77,7 +59,7 @@ func TestAccMorpheusTaskEmailExampleOk(t *testing.T) {
 
 	name := acctest.RandomWithPrefix(t.Name())
 
-	resourceConfig, err := RenderTaskEmailConfig(t, map[string]string{"Name": name})
+	resourceConfig, err := RenderMorpheusTaskEmailConfig(t, name, nil)
 	if err != nil {
 		t.Fatal(err)
 	}
