+++ internal/sdkv2/morpheus/resources/task/morpheus_task_email_resource_url.tf.tmpl
@@ -1,4 +1,4 @@
-resource "hpe_morpheus_task_email" "{{.Name}}" {
+resource "hpe_morpheus_task_email" "tfexample_email_url" {
   name                        = "{{.Name}}"
   code                        = "{{.Code}}"
   labels                      = {{.Labels}}
